Added logging function name on ServiceBus exception.
