+++ sdk/servicebus/Microsoft.Azure.WebJobs.Extensions.ServiceBus/tests/Config/ServiceBusOptionsTests.cs
@@ -44,50 +44,6 @@ namespace Microsoft.Azure.WebJobs.ServiceBus.UnitTests.Config
         }
 
         [Test]
-        public void LogExceptionReceivedEvent_NonTransientEvent_LoggedAsError()
-        {
-            var ex = new ServiceBusException(isTransient: false, message: "message");
-            Assert.False(ex.IsTransient);
-            ProcessErrorEventArgs e = new ProcessErrorEventArgs(ex, ServiceBusErrorSource.Abandon, "TestEndpoint", "TestEntity", CancellationToken.None);
-            ServiceBusExtensionConfigProvider.LogExceptionReceivedEvent(e, _loggerFactory);
-
-            var expectedMessage = $"Message processing error (Action=Abandon, EntityPath=TestEntity, Endpoint=TestEndpoint)";
-            var logMessage = _loggerProvider.GetAllLogMessages().Single();
-            Assert.AreEqual(LogLevel.Error, logMessage.Level);
-            Assert.AreSame(ex, logMessage.Exception);
-            Assert.AreEqual(expectedMessage, logMessage.FormattedMessage);
-        }
-
-        [Test]
-        public void LogExceptionReceivedEvent_TransientEvent_LoggedAsInformation()
-        {
-            var ex = new ServiceBusException(message: "message", isTransient: true);
-            Assert.True(ex.IsTransient);
-            ProcessErrorEventArgs e = new ProcessErrorEventArgs(ex, ServiceBusErrorSource.Receive, "TestEndpoint", "TestEntity", CancellationToken.None);
-            ServiceBusExtensionConfigProvider.LogExceptionReceivedEvent(e, _loggerFactory);
-
-            var expectedMessage = $"Message processing error (Action=Receive, EntityPath=TestEntity, Endpoint=TestEndpoint)";
-            var logMessage = _loggerProvider.GetAllLogMessages().Single();
-            Assert.AreEqual(LogLevel.Information, logMessage.Level);
-            Assert.AreSame(ex, logMessage.Exception);
-            Assert.AreEqual(expectedMessage, logMessage.FormattedMessage);
-        }
-
-        [Test]
-        public void LogExceptionReceivedEvent_NonMessagingException_LoggedAsError()
-        {
-            var ex = new MissingMethodException("What method??");
-            ProcessErrorEventArgs e = new ProcessErrorEventArgs(ex, ServiceBusErrorSource.Complete, "TestEndpoint", "TestEntity", CancellationToken.None);
-            ServiceBusExtensionConfigProvider.LogExceptionReceivedEvent(e, _loggerFactory);
-
-            var expectedMessage = $"Message processing error (Action=Complete, EntityPath=TestEntity, Endpoint=TestEndpoint)";
-            var logMessage = _loggerProvider.GetAllLogMessages().Single();
-            Assert.AreEqual(LogLevel.Error, logMessage.Level);
-            Assert.AreSame(ex, logMessage.Exception);
-            Assert.AreEqual(expectedMessage, logMessage.FormattedMessage);
-        }
-
-        [Test]
         public void ToProcessorOptions_ReturnsExpectedValue()
         {
             ServiceBusOptions sbOptions = new ServiceBusOptions
